InstanceExport: Adding ServerHealthCheck and retry mechanism
Adding a new function ServerHealthCheck to do a get to the /en/healthcheck to check if the site is online. For the cases the site is off or deploying we will receive a 503 response and we will wait till the server is back to continue the process, avoiding to try to download possibly a lot of data that may be available when site is back.